Move proto to protos subdir to fix  sync to arcadia
