+++ ydb/core/driver_lib/run/ya.make
@@ -192,7 +192,7 @@ PEERDIR(
 IF (OS_LINUX)
     PEERDIR(
         ydb/core/nbs/cloud/blockstore/bootstrap
-        ydb/core/nbs/cloud/blockstore/config
+        ydb/core/nbs/cloud/blockstore/config/protos
         ydb/core/nbs/cloud/blockstore/libs/storage/ss_proxy
         ydb/core/nbs/cloud/blockstore/libs/storage/volume
