+++ ydb/core/nbs/cloud/blockstore/config/protos/ya.make
@@ -1,0 +1,10 @@
+PROTO_LIBRARY()
+
+INCLUDE_TAGS(GO_PROTO)
+EXCLUDE_TAGS(JAVA_PROTO)
+
+SRCS(
+    storage.proto
+)
+
+END()
