+++ ydb/core/nbs/cloud/blockstore/config/ya.make
@@ -1,10 +1,2 @@
-PROTO_LIBRARY()
-
-INCLUDE_TAGS(GO_PROTO)
-EXCLUDE_TAGS(JAVA_PROTO)
-
-SRCS(
-    storage.proto
+RECURSE(
 )
-
-END()
