+++ ydb/core/nbs/cloud/blockstore/libs/storage/core/ya.make
@@ -7,7 +7,7 @@ SRCS(
 )
 
 PEERDIR(
-    ydb/core/nbs/cloud/blockstore/config
+    ydb/core/nbs/cloud/blockstore/config/protos
     ydb/core/nbs/cloud/blockstore/libs/storage/model
     ydb/core/nbs/cloud/storage/core/libs/common
     ydb/core/nbs/cloud/storage/core/protos
