+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/partition_direct.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ydb/core/nbs/cloud/blockstore/config/storage.pb.h>
+#include <ydb/core/nbs/cloud/blockstore/config/protos/storage.pb.h>
 
 #include <ydb/core/base/blobstorage.h>
 #include <ydb/core/engine/minikql/flat_local_tx_factory.h>
