+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/ya.make
@@ -15,7 +15,7 @@ SRCS(
 
 PEERDIR(
     ydb/core/nbs/cloud/blockstore/bootstrap
-    ydb/core/nbs/cloud/blockstore/config
+    ydb/core/nbs/cloud/blockstore/config/protos
     ydb/core/nbs/cloud/blockstore/libs/storage/api
     ydb/core/nbs/cloud/blockstore/libs/storage/storage_transport
     ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/protos
