+++ ydb/core/nbs/cloud/blockstore/ya.make
@@ -1,5 +1,6 @@
 RECURSE(
     bootstrap
+    config
     libs
     public
     tools
